+++ maxcompute-20220104/core/Models/SumStorageMetricsByDateRequest.cs
@@ -1,0 +1,58 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.MaxCompute20220104.Models
+{
+    public class SumStorageMetricsByDateRequest : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1718590596556</para>
+        /// </summary>
+        [NameInMap("endDate")]
+        [Validation(Required=false)]
+        public long? EndDate { get; set; }
+
+        [NameInMap("projectNames")]
+        [Validation(Required=false)]
+        public List<string> ProjectNames { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>cn-hangzhou</para>
+        /// </summary>
+        [NameInMap("region")]
+        [Validation(Required=false)]
+        public string Region { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>1715393576201</para>
+        /// </summary>
+        [NameInMap("startDate")]
+        [Validation(Required=false)]
+        public long? StartDate { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>PROJECT</para>
+        /// </summary>
+        [NameInMap("statsType")]
+        [Validation(Required=false)]
+        public string StatsType { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>12345</para>
+        /// </summary>
+        [NameInMap("userId")]
+        [Validation(Required=false)]
+        public string UserId { get; set; }
+
+    }
+
+}