+++ maxcompute-20220104/core/Models/SumStorageMetricsByDateResponseBody.cs
@@ -1,0 +1,114 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.MaxCompute20220104.Models
+{
+    public class SumStorageMetricsByDateResponseBody : TeaModel {
+        [NameInMap("data")]
+        [Validation(Required=false)]
+        public List<SumStorageMetricsByDateResponseBodyData> Data { get; set; }
+        public class SumStorageMetricsByDateResponseBodyData : TeaModel {
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>20250719</para>
+            /// </summary>
+            [NameInMap("dateTime")]
+            [Validation(Required=false)]
+            public string DateTime { get; set; }
+
+            [NameInMap("itemStorageMetrics")]
+            [Validation(Required=false)]
+            public List<SumStorageMetricsByDateResponseBodyDataItemStorageMetrics> ItemStorageMetrics { get; set; }
+            public class SumStorageMetricsByDateResponseBodyDataItemStorageMetrics : TeaModel {
+                /// <summary>
+                /// <b>Example:</b>
+                /// <para>prj</para>
+                /// </summary>
+                [NameInMap("itemName")]
+                [Validation(Required=false)]
+                public string ItemName { get; set; }
+
+                /// <summary>
+                /// <b>Example:</b>
+                /// <para>91.22</para>
+                /// </summary>
+                [NameInMap("percentage")]
+                [Validation(Required=false)]
+                public double? Percentage { get; set; }
+
+                /// <summary>
+                /// <b>Example:</b>
+                /// <para>300.560392</para>
+                /// </summary>
+                [NameInMap("usage")]
+                [Validation(Required=false)]
+                public string Usage { get; set; }
+
+            }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>Storage</para>
+            /// </summary>
+            [NameInMap("storageType")]
+            [Validation(Required=false)]
+            public string StorageType { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>GB</para>
+            /// </summary>
+            [NameInMap("unit")]
+            [Validation(Required=false)]
+            public string Unit { get; set; }
+
+            /// <summary>
+            /// <b>Example:</b>
+            /// <para>329.503338</para>
+            /// </summary>
+            [NameInMap("usage")]
+            [Validation(Required=false)]
+            public string Usage { get; set; }
+
+        }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>success</para>
+        /// </summary>
+        [NameInMap("errorCode")]
+        [Validation(Required=false)]
+        public string ErrorCode { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>this quota is not exist.</para>
+        /// </summary>
+        [NameInMap("errorMsg")]
+        [Validation(Required=false)]
+        public string ErrorMsg { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>200</para>
+        /// </summary>
+        [NameInMap("httpCode")]
+        [Validation(Required=false)]
+        public int? HttpCode { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>0abb781a17411408145995819e0dae</para>
+        /// </summary>
+        [NameInMap("requestId")]
+        [Validation(Required=false)]
+        public string RequestId { get; set; }
+
+    }
+
+}